requirements.txt
transformers
pillow
python-dotenv
numpy
fastapi
uvicorn
httpx
